refactor:代码优化

+++ youlai-admin/admin-api/src/main/java/com/youlai/admin/pojo/SysRoleResource.java
@@ -6,6 +6,6 @@ import lombok.experimental.Accessors;
 @Data
 @Accessors(chain = true)
 public class SysRoleResource {
-    private Integer roleId;
-    private Integer resourceId;
+    private Long roleId;
+    private Long resourceId;
 }
